+++ src/index.js
@@ -20,3 +20,5 @@ module.exports = {
 	},
 	defaultOptions,
 };
+
+// https://www.youtube.com/watch?v=Yoj-VLuCZvA