+++ test/test.js
@@ -28,6 +28,7 @@ const { getVideo, getPlaylist, search } = require('../')
 
 	console.log('Works well :D');
 
+	return;
 	fs.writeFileSync(
 		fs.existsSync('./test') ? './test/results.json' : './results.json',
 		JSON.stringify(data, null, '\t')